move new gantt chart component
